move state upgrade related functions to own file (#684)
